+++ ethereum/json-types/src/main/java/tech/pegasys/teku/ethereum/json/types/beacon/GetGenesisApiData.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.ethereum.json.types.wrappers;
+package tech.pegasys.teku.ethereum.json.types.beacon;
 
 import com.google.common.base.MoreObjects;
 import java.util.Objects;
